% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btergm_count_bootstrap.R,
%   R/btergm_count_methods.R
\name{coef.ergmCntMPLE}
\alias{coef.ergmCntMPLE}
\alias{btergm_count_methods}
\alias{coef.btergm_count}
\alias{print.btergm_count}
\alias{nobs.btergm_count}
\alias{confint.btergm_count}
\alias{summary.btergm_count}
\alias{print.summary.btergm_count}
\alias{coef,btergm_count-method}
\alias{nobs,btergm_count-method}
\alias{confint,btergm_count-method}
\alias{summary,btergm_count-method}
\title{Methods for btergm_count objects}
\usage{
\method{coef}{ergmCntMPLE}(object, ...)

\method{coef}{btergm_count}(object, ...)

\method{print}{btergm_count}(x, digits = max(3, getOption("digits") - 3), ...)

\method{nobs}{btergm_count}(object, ...)

\method{confint}{btergm_count}(object, parm, level = 0.95, type = c("perc", "basic", "norm"), ...)

\method{summary}{btergm_count}(
  object,
  level = 0.95,
  type = c("perc", "basic", "norm"),
  include.mple.se = TRUE,
  ...
)

\method{print}{summary.btergm_count}(x, digits = max(3, getOption("digits") - 3), ...)
}
\description{
Basic methods for extracting, printing, summarizing, and constructing
confidence intervals for fitted count-valued bootstrapped TERGM objects.
}
\keyword{internal}
