% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btergm_ergmMPLE_sampled.R
\name{.btergm_arr_from_coo}
\alias{.btergm_arr_from_coo}
\title{Retrieve arr_from_coo from the available statnet namespace}
\usage{
.btergm_arr_from_coo()
}
\value{
The internal \code{arr_from_coo()} helper.
}
\description{
Retrieve arr_from_coo from the available statnet namespace
}
\keyword{internal}
