% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btergm_ergmMPLE_sampled.R
\name{.btergm_nonsimp_update_formula}
\alias{.btergm_nonsimp_update_formula}
\title{Retrieve a formula updater compatible with the installed statnet version}
\usage{
.btergm_nonsimp_update_formula()
}
\value{
A function for updating ERGM formulas.
}
\description{
Retrieve a formula updater compatible with the installed statnet version
}
\keyword{internal}
